clean up & update readmes
