+++ asset/README.md
@@ -1,0 +1,7 @@
+# asset
+this directory contains media assets such as:
+- images
+- sprite sheets for animations
+- fonts
+- sound files
+- etc
