--- asset/misc/Buy GGBotNet a Coffee-543f.url
@@ -1,6 +1,0 @@
-[InternetShortcut]
-URL=https://ko-fi.com/ggbotnet
-IDList=
-HotKey=0
-IconFile=C:\Users\SP700\AppData\Local\Mozilla\Firefox\Profiles\tlgohe1j.default-release-1686839021344\shortcutCache\c35agND88zEjVb56xx6xbw==.ico
-IconIndex=0
