--- asset/misc/GGBotNet - itch.io-3086.url
@@ -1,6 +1,0 @@
-[InternetShortcut]
-URL=https://ggbot.itch.io/
-IDList=
-HotKey=0
-IconFile=C:\Users\SP700\AppData\Local\Mozilla\Firefox\Profiles\tlgohe1j.default-release-1686839021344\shortcutCache\CqQO97+aqBcylB94dwTjcg==.ico
-IconIndex=0
