+++ data/README.md
@@ -1,0 +1,2 @@
+# data
+this directory contains `.toml` files to make it easier to manage various constants and variables.
