+++ planetclicker/__init__.py
@@ -2,5 +2,3 @@
 Planet Clicker Module
 =====================
 """
-
-# __all__ = ["data", "asset", "model", "scene", "sprite", "ui"]
