+++ planetclicker/data/namespace.py
@@ -1,49 +1,7 @@
-import os
-from collections import defaultdict
-from tomlkit.api import string
-from dataclasses import dataclass
 from types import SimpleNamespace
 
-import pygame
 import tomlkit
-
-from planetclicker.asset.font import FontsParser
-
-
-@dataclass
-class TOMLReader:
-    path: str
-    # def read(self, filename: str):
-    #     s
-
-    def get(self, key: str):
-        return self.data.get(key)
-
-    def gets(self, *keys: str):
-        cfg = self.data
-        for key in keys:
-            cfg = cfg.get(key)
-        return cfg
-
-    @property
-    def data(self):
-        with open(self.path) as fp:
-            doc = tomlkit.load(fp)
-            # doc.add("path", string(self.path))
-            return doc
-
-
-@dataclass
-class FontReader(TOMLReader):
-    path: str
-
-    @property
-    def fonts(self):
-        fonts = {}
-        for label, props in self.data.items():
-            path = os.path.join("asset", "font", "free", props["name"])
-            fonts[label] = pygame.font.Font(path, props["size"])
-        return fonts
+from planetclicker.data.loader import FontReader, TOMLReader
 
 
 class Namespace(SimpleNamespace):
@@ -67,9 +25,6 @@ class Namespace(SimpleNamespace):
     def data(self):
         with open(self.path) as fp:
             tomlkit.load(fp)
-
-
-font_parser = FontsParser()
 
 
 class Game(Namespace):
