--- planetclicker/scene/__init__.py
@@ -1,6 +1,0 @@
-from dataclasses import dataclass
-from typing import Optional
-
-import pygame
-
-__all__ = ["scene"]
